refactor: replace AtomStoreContext
