+++ src/core/atomStore/AtomStoreContext.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+import { IAtomStore, createAtomStore } from '..';
+
+export const AtomStoreContext = createContext<IAtomStore>(createAtomStore());
